Переделала запсиь/чтение в бинарник св-ва alnAt в break (в Run)
git-svn-id: svn://192.168.3.15/activex/AVS/Sources/TeamlabOffice/trunk/OfficeWeb@64163 954022d7-b5bf-4e40-9824-e11837661b57

+++ Word/Math/borderBox.js
@@ -28,7 +28,7 @@ CMathBreak.prototype.Write_ToBinary = function(Writer)
     if(this.alnAt !== undefined)
     {
         Writer.WriteBool(false);
-        Writer.WriteLong(this.alnAt);
+        Writer.WriteByte(this.alnAt);
     }
     else
     {
@@ -39,7 +39,7 @@ CMathBreak.prototype.Read_FromBinary = function(Reader)
 {
     if(Reader.GetBool() == false)
     {
-        this.alnAt = Reader.GetLong();
+        this.alnAt = Reader.GetByte();
     }
     else
     {
